requirements.txt
# === Core Cryptography ===
pycryptodome==3.21.0

# === Networking and API (for later phases) ===
flask==3.0.3
requests==2.32.3

# === Blockchain Integration (Phase 4–5) ===
web3==6.15.1

# === Database (Optional, for user authentication) ===
sqlalchemy==2.0.36

# === Utility Tools ===
colorama==0.4.6   # for colorful terminal logs
tqdm==4.66.5      # for progress bars

# === Testing and Linting (Optional) ===
pytest==8.3.3
flake8==7.1.1

# === UI ===
streamlit